pacemaker: Ignore node CIB objects when looking for a definition
In some very rare cases, we can get "crm configure show $foo" return
more than one CIB objects. This seems to happen for remote nodes that
have attributes.

$ crm configure show remote-d52-54-77-77-77-04
node remote-d52-54-77-77-77-04:remote \
        attributes OpenStack-role=compute
primitive remote-d52-54-77-77-77-04 ocf:pacemaker:remote \
        params reconnect_interval=60s server=d52-54-77-77-77-04.h9.cloud.suse.de \
        op monitor interval=20s \
        op start timeout=60s interval=0 \
        op stop timeout=60s interval=0

Unfortunately, when we parse the output of the command to create a CIB
object, we don't necessarily know the type of the object we're looking
for.

To work around this, we simply ignore node objects.