crowbar-pacemaker: Make haproxy active/active
This requires a few changes:

 - enabling binding to non-local addresses (sysctl setting), as haproxy
   may listen to a VIP that is not on this node
 - removing the g-haproxy group that was grouping the VIP and haproxy
   together, to provide ordering and colocation constraints (which are
   not required anymore)

We do not need the ordering constraint anymore, thanks to the non-local
bind setting, which means the VIP can be started after haproxy.

We do not need th colocation constraint anymore, because haproxy runs
everywhere anyhow.

TODO: We would probably still need a way to say that a VIP should not
run on a node where haproxy is stopped.